Working but it doesnt have the pager component
