+++ src/components/category/category.jsx
@@ -6,7 +6,10 @@ import { classnames } from "../../helpers";
 import * as styles from "./category.module.scss";
 
 const Category = ({ data, className }) => (
+  <>
+  {console.log(data)}
   <p className={classnames(styles.categoryWrapper, className)}>{data}</p>
+  </>
 );
 
 Category.propTypes = {
