+++ src/templates/categories.module.scss
